Add OoyalaCachePlugin
The OoyalaCachePlugin is used to easily provide a cache backend configuration. There are some defaults (like SkipRevalidation) that are needed to actually cache ooyala responses because of issues getting back valid Cache headers in the response.

Custom "can cache" strategies are provided as defaults to get around the lack of Cache response headers.